chore: Rename files in _shared folder
